multi-agent-chatbot/requirements.txt
langgraph
langchain-core
langchain-google-genai
python-dotenv
langchain-google-community[gmail]
langchain-google-vertexai
ipython
pymupdf
sentence_transformers
python-docx
faiss-cpu
langchain-cohere
